+++ pynrc/maths/fast_poly.py
