+++ pynrc/maths/image_manip.py
