+++ pynrc/maths/robust.py
