+++ pynrc/notebooks/2mass_ks.txt
