+++ pynrc/notebooks/coronagraph_properties/coronagraph_references.xlsx
