+++ pynrc/notebooks/coronagraph_properties/references_gamma.pdf
