+++ pynrc/notebooks/coronagraph_properties/references_magnitudes.pdf
