+++ pynrc/notebooks/coronagraph_properties/references_sptype.pdf
