+++ pynrc/notebooks/example_disk.fits
