+++ pynrc/notebooks/grism_properties/2mass_ks.txt
