+++ pynrc/notebooks/grism_properties/GRISM0_ModA.tbl
