+++ pynrc/notebooks/grism_properties/GRISM0_ModB.tbl
