+++ pynrc/notebooks/grism_properties/GRISM90_ModA.tbl
