+++ pynrc/notebooks/grism_properties/GRISM90_ModB.tbl
