+++ pynrc/notebooks/grism_properties/Grism_properties.xlsx
