+++ pynrc/notebooks/hd142527_hband.fits
