+++ pynrc/reduce/ref_pixels.py
