+++ pynrc/speckle_noise.py
