+++ pynrc/version.py
@@ -1,3 +1,3 @@
-__version__ = u'0.5.0'
+__version__ = u'0.6.0'
 #__author__ = 'Jarron Leisenring'
 #__email__  = 'jarronl@email.arizona.edu'
